# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "version": "1.0.0",
    "title": "hw2-mongodb",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "description": "API for a notebook application using MongoDB"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    },
    {
      "url": "https://backend-taskpro.onrender.com"
    }
  ],
...
  "paths": {
...
    "/cards/:columnId": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get list of cards",
        "operationId": "GetCards",
        "description": "Get list of cards with this endpoint",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully found cards!",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found cards!"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "example": {
                          "type": "object",
                          "required": [
                            "title",
                            "description",
                            "priority",
                            "date"
                          ],
                          "properties": {
                            "_id": {
                              "description": "Card id",
                              "type": "string",
                              "example": "67280fd53bdc839326b08e78"
                            },
                            "columnId": {
                              "description": "Card columnId",
                              "type": "string",
                              "example": "67280ee53bdc839326b08e24"
                            },
                            "boardId": {
                              "description": "Card boardId",
                              "type": "string",
                              "example": "67280e853bdc839326b08e0f"
                            },
                            "title": {
                              "description": "Card title",
                              "type": "string",
                              "example": "Materials"
                            },
                            "description": {
                              "description": "Card description",
                              "type": "string",
                              "example": "Link of materials http://materials.link.com"
                            },
                            "priority": {
                              "description": "Card priority",
                              "type": "string",
                              "enum": [
                                "low",
                                "medium",
                                "high",
                                "without"
                              ],
                              "example": "without"
                            },
                            "date": {
                              "description": "Card date",
                              "type": "string",
                              "example": "2024-10-12T05:46:48.395Z"
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      },
      "post": {
        "tags": [
          "cards"
        ],
        "summary": "Post card",
        "operationId": "createCard",
        "description": "create a card.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "columnId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "title",
                  "description",
                  "priority",
                  "date"
                ],
                "properties": {
                  "title": {
                    "description": "Column title",
                    "type": "string",
                    "example": "Task#1"
                  },
                  "description": {
                    "description": "Column title",
                    "type": "string",
                    "example": "Task#1 - fill list task"
                  },
                  "priority": {
                    "type": "string",
                    "enum": [
                      "low",
                      "medium",
                      "hight"
                    ],
                    "example": "low"
                  },
                  "date": {
                    "type": "string",
                    "example": "2024-11-06T14:33:38.468Z"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Create a card",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Column!"
                    },
                    "data": {
                      "type": "string",
                      "example": {
                        "type": "object",
                        "required": [
                          "title",
                          "description",
                          "priority",
                          "date"
                        ],
                        "properties": {
                          "_id": {
                            "description": "Card id",
                            "type": "string",
                            "example": "67280fd53bdc839326b08e78"
                          },
                          "columnId": {
                            "description": "Card columnId",
                            "type": "string",
                            "example": "67280ee53bdc839326b08e24"
                          },
                          "boardId": {
                            "description": "Card boardId",
                            "type": "string",
                            "example": "67280e853bdc839326b08e0f"
                          },
                          "title": {
                            "description": "Card title",
                            "type": "string",
                            "example": "Materials"
                          },
                          "description": {
                            "description": "Card description",
                            "type": "string",
                            "example": "Link of materials http://materials.link.com"
                          },
                          "priority": {
                            "description": "Card priority",
                            "type": "string",
                            "enum": [
                              "low",
                              "medium",
                              "high",
                              "without"
                            ],
                            "example": "without"
                          },
                          "date": {
                            "description": "Card date",
                            "type": "string",
                            "example": "2024-10-12T05:46:48.395Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    },
    "/cards/:cardId": {
      "get": {
        "tags": [
          "cards"
        ],
        "summary": "Get card by id",
        "operationId": "getCard",
        "description": "Get a card data.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cardId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "672a8db61a46f387714eac47"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get card by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully found Column!"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "type": "object",
                        "required": [
                          "title",
                          "description",
                          "priority",
                          "date"
                        ],
                        "properties": {
                          "_id": {
                            "description": "Card id",
                            "type": "string",
                            "example": "67280fd53bdc839326b08e78"
                          },
                          "columnId": {
                            "description": "Card columnId",
                            "type": "string",
                            "example": "67280ee53bdc839326b08e24"
                          },
                          "boardId": {
                            "description": "Card boardId",
                            "type": "string",
                            "example": "67280e853bdc839326b08e0f"
                          },
                          "title": {
                            "description": "Card title",
                            "type": "string",
                            "example": "Materials"
                          },
                          "description": {
                            "description": "Card description",
                            "type": "string",
                            "example": "Link of materials http://materials.link.com"
                          },
                          "priority": {
                            "description": "Card priority",
                            "type": "string",
                            "enum": [
                              "low",
                              "medium",
                              "high",
                              "without"
                            ],
                            "example": "without"
                          },
                          "date": {
                            "description": "Card date",
                            "type": "string",
                            "example": "2024-10-12T05:46:48.395Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "patch": {
        "tags": [
          "cards"
        ],
        "summary": "Update card by id",
        "operationId": "updateCard",
        "description": "Update a card data.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cardId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "672a8db61a46f387714eac47"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Update card by id",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "status",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "status": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Successfully update Card!"
                    },
                    "data": {
                      "type": "object",
                      "example": {
                        "type": "object",
                        "required": [
                          "title",
                          "description",
                          "priority",
                          "date"
                        ],
                        "properties": {
                          "_id": {
                            "description": "Card id",
                            "type": "string",
                            "example": "67280fd53bdc839326b08e78"
                          },
                          "columnId": {
                            "description": "Card columnId",
                            "type": "string",
                            "example": "67280ee53bdc839326b08e24"
                          },
                          "boardId": {
                            "description": "Card boardId",
                            "type": "string",
                            "example": "67280e853bdc839326b08e0f"
                          },
                          "title": {
                            "description": "Card title",
                            "type": "string",
                            "example": "Materials"
                          },
                          "description": {
                            "description": "Card description",
                            "type": "string",
                            "example": "Link of materials http://materials.link.com"
                          },
                          "priority": {
                            "description": "Card priority",
                            "type": "string",
                            "enum": [
                              "low",
                              "medium",
                              "high",
                              "without"
                            ],
                            "example": "without"
                          },
                          "date": {
                            "description": "Card date",
                            "type": "string",
                            "example": "2024-10-12T05:46:48.395Z"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/401"
          }
        }
      },
      "delete": {
        "tags": [
          "cards"
        ],
        "summary": "Delete card by id",
        "operationId": "deleteCard",
        "description": "Delete a card by id.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "cardId",
            "required": true,
            "schema": {
              "type": "string",
              "example": "65ca67e7ae7f10c88b598384"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "404": {
            "$ref": "#/components/responses/404"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
...
    "responses": {
      "401": {
        "description": "UnauthorizedError",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 401
                },
                "message": {
                  "type": "string",
                  "example": "UnauthorizedError"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Invalid password"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "404": {
        "description": "Not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": [
                "status",
                "message"
              ],
              "properties": {
                "status": {
                  "type": "integer",
                  "example": 404
                },
                "message": {
                  "type": "string",
                  "example": "User not found"
                }
              }
            }
          }
        }
      },
...
    }
  }
}